Cleanup/cmake (#61)
* explicitly request to use installed libpotassco

With this commit there is a separate option to request whether the
installed libpotassco is used. The reason for this change is that once
the library is installed with the install target, then the current setup
will switch to using the installed library and no longer use the
submodule. Furthermore, this allows clingo to explicitly request not to
use the system installation because this will not work (it would be a
nice future project to also support this use case). The commit also
makes it the default to use the submodule because this is the most
common use case.

* more flexible/modern installation

This commit adds an option to install using a versioned or unversioned
layout. Furthermore it uses cmake's package config helpers to simplify
config file generation.